Tweak Sodium donation button text
Now reads "Support CaffeineMC!" rather than "Buy us a coffee!"

This change was made in an effort to make it clear that donations through the button do not go to me (AeiouEnigma), as the use of "us" might have implied. Simply moving the donation button wouldn't have felt like an appropriate thing to do.